fix: conditionaly invioke clear folders step in plugin

+++ apps/molgenis-components/docs-plugin.js
@@ -38,8 +38,11 @@ module.exports = function () {
   };
 
   const buildStart = () => {
-    fs.rmSync("./showCase/", { recursive: true, force: true });
-    fs.rmSync("./gen-docs/", { recursive: true, force: true });
+    if (process.env.CLEAR_GEN_FOLDERS === 'on') {
+      console.log("CLEAR_GEN_FOLDERS is set to 'on', clearing folders");
+      fs.rmSync("./showCase/", { recursive: true, force: true });
+      fs.rmSync("./gen-docs/", { recursive: true, force: true });
+    }
   };
 
   return { name: "docs-plugin", buildStart, transform };
